Rename layers folder to modules
